+++ src/domains/login/pages/Login.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Link } from "react-router-dom";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faHouse } from "@fortawesome/free-solid-svg-icons";
-import LoginSocialBtn from "../components/Login/LoginSocialBtn";
+import LoginSocialBtn from "../components/LoginSocialBtn";
 
 const Login = () => {
     return (
